refactor: auth 컨트롤러를 통합
