+++ src/main/java/io/cavia/trader/module/auth/controller/AuthViewController.java
@@ -4,7 +4,18 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class PasswordViewController {
+public class AuthViewController {
+
+
+    @GetMapping("/login")
+    public String showLoginForm() {
+        return "member/login";
+    }
+
+    @GetMapping("login-checker")
+    public String showLoginChecker() {
+        return "member/login-checker";
+    }
 
     @GetMapping("/forgot-password")
     public String showPasswordForm() {
